# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Restful API for web app to buy air tickets
  version: 1.1.0
servers:
  - url: https://api.airtickets.com
paths:
...
  /users:
    post:
      summary: Create user
      tags:
        - Users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: Successful response with created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '409':
          description: User exists
        '400':
          description: Incorrect input data
  /users/{email}:
    get:
      summary: Get user by id
      tags:
        - Users
      operationId: getUserById
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response with user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Incorrect input data
        '404':
          description: User is not in DB
  /users/{userId}:
    patch:
      summary: Modify user by id
      tags:
        - Users
      operationId: changeUserById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      parameters: 
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with modified user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
    delete:
      summary: Delete user by id
      tags:
        - Users
      operationId: deleteUserById
      parameters: 
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
components:
  schemas:
...
    User:
      type: object
      required:
        - userId
        - role
        - email
        - password
        - regDate
      properties:
        userId:
          readOnly: true
          type: integer
          example: 3
        role:
          readOnly: true
          type: string
          example: клиент
          enum:
            - клиент
            - модератор
            - админ
        email:
          type: string
          example: email@mail.domain
        password:
          writeOnly: true
          type: string
          example: aksdsaj3313jkj1k23h1jk3b21ji3b21b
        firstName:
          type: string
          example: Vladyslav
        lastName:
          type: string
          example: Marchenko
        regDate:
          type: string
          readOnly: true
          format: 2023-08-05T15:00:00:02
          example: 2023-08-05T15:00:00:02
...